+++ src/api/index.ts
@@ -1,20 +1,21 @@
-export * from './apps';
-export * from './member';
-export * from './item';
-export * from './membership';
-export * from './authentication';
-export * from './itemTag';
-export * from './itemLogin';
-export * from './itemFlag';
-export * from './chat';
-export * from './mentions';
-export * from './category';
-export * from './search';
-export * from './itemExport';
-export * from './itemLike';
-export * from './itemValidation';
 export * from './action';
+export * from './apps';
+export * from './authentication';
+export * from './category';
+export * from './chat';
+export * from './etherpad';
 export * from './invitation';
-export * from './subscription';
-export * from './itemPublish';
+export * from './item';
+export * from './itemExport';
 export * from './itemFavorite';
+export * from './itemFlag';
+export * from './itemLike';
+export * from './itemLogin';
+export * from './itemPublish';
+export * from './itemTag';
+export * from './itemValidation';
+export * from './member';
+export * from './membership';
+export * from './mentions';
+export * from './search';
+export * from './subscription';
