+++ src/api/item.ts
@@ -1,10 +1,4 @@
-import {
-  DiscriminatedItem,
-  EtherpadItemType,
-  Item,
-  ResultOf,
-  UUID,
-} from '@graasp/sdk';
+import { DiscriminatedItem, Item, ResultOf, UUID } from '@graasp/sdk';
 
 import { DEFAULT_THUMBNAIL_SIZE } from '../config/constants';
 import { QueryClientConfig } from '../types';
@@ -20,13 +14,11 @@ import {
   buildDownloadItemThumbnailRoute,
   buildEditItemRoute,
   buildGetChildrenRoute,
-  buildGetEtherpadRoute,
   buildGetItemDescendants,
   buildGetItemParents,
   buildGetItemRoute,
   buildGetItemsRoute,
   buildMoveItemsRoute,
-  buildPostEtherpadRoute,
   buildPostItemRoute,
   buildRecycleItemsRoute,
   buildRestoreItemsRoute,
@@ -252,31 +244,4 @@ export const downloadItemThumbnailUrl = async (
         replyUrl: true,
       })}`,
     )
-    .then(({ data }) => data);
-
-export const postEtherpad = async (
-  {
-    name,
-    parentId,
-  }: Pick<Item, 'name'> & {
-    parentId?: UUID;
-  },
-  { API_HOST }: QueryClientConfig,
-): Promise<EtherpadItemType> =>
-  verifyAuthentication(() =>
-    axios
-      .post(`${API_HOST}/${buildPostEtherpadRoute(parentId)}`, {
-        name: name.trim(),
-      })
-      .then(({ data }) => data),
-  );
-
-export const getEtherpad = (
-  { itemId, mode }: { itemId: UUID; mode: 'read' | 'write' },
-  { API_HOST }: QueryClientConfig,
-) =>
-  axios
-    .get(`${API_HOST}/${buildGetEtherpadRoute(itemId)}`, {
-      params: { mode },
-    })
     .then(({ data }) => data);
